corrected issue

+++ clean_rpath.sh
@@ -10,7 +10,11 @@ for bin in $(find bin -mindepth 1); do
   for lib_path in $(otool -l $bin | grep -A2 LC_LOAD_DYLIB | grep dylib | grep name | awk '{print $2}' | grep -v "@.*path" | grep -v "^/usr/lib/" | grep -v "^/usr/local/lib/" | grep -v "^/System"); do
     lib_name_depend=$(echo $lib_path | awk -F'/' '{print $(NF)}')
     install_name_tool -change $lib_path @rpath/${lib_name_depend} $bin
-    cp $lib_path ../lib
+    if [ -f "$lib_path" ]; then
+      if [ ! -f "../lib/${lib_name_depend}" ]; then
+        cp $lib_path ../lib
+      fi
+    fi
   done
 done
 
@@ -18,7 +22,9 @@ for lib in $(find lib -mindepth 1 -name "*.dylib"); do
   install_name_tool -id @rpath/$(basename $lib) $lib
   for lib_path in $(otool -l $lib | grep -A2 LC_LOAD_DYLIB | grep dylib | grep name | awk '{print $2}' | grep -v "@.*path" | grep -v "^/usr/lib/" | grep -v "^/usr/local/lib/" | grep -v "^/System"); do
     if [ -f "$lib_path" ]; then
-      cp $lib_path lib
+      if [ ! -f "../lib/${lib_name_depend}" ]; then
+        cp $lib_path lib
+      fi
       if [ -f "lib/$(basename $lib_path)" ]; then
         for lib_path_dylib in $(otool -l lib/$(basename $lib_path) | grep -A2 LC_LOAD_DYLIB | grep dylib | grep name | awk '{print $2}' | grep -v "@.*path" | grep -v "^/usr/lib/" | grep -v "^/usr/local/lib/" | grep -v "^/System"); do
           install_name_tool -change $lib_path_dylib @rpath/$(echo $lib_path_dylib | awk -F'/' '{print $(NF)}') lib/$(basename $lib_path)
